v2.0.19/v2.1.3
problem: when the `-lt` flag was used with libtorrent `v1.2` like `-lt v1.2.20` boost was not being defaulted `boost-1.86.0` as it does when setting the env equivalent causing a build error when libtorrent was built.

fix: the flag now properly checks and sets boost to `boost-1.86.0` when libtorrent `v1.2` is being built.