fix: revisions and release
Sometimes when the workflow gets the dependency.json it retrieves, from best I can determine, an older version of the file with outdated info. It's not failing to retrieve the file, which I initially thought was the case, it's getting the file, just not the current version of it. I've noticed this happen with more frequency lately.

You assume that this could happen when running the tests in quick succession as the file might not have propagated across their network. The problem is this is happening with release assets that have been sitting there for weeks. I file I know has a revision of 2 yet the file it reads has 1.

This must be some Github caching or system to work around action issues. In the testing for this changes my matrix would return something like all correct or 30% previous versions.

So this change is a band aid.

When github implements immutable tags i'll probably move to using revision appended tags.